fix(ui): refactor and moved deletion dialog and logic into editor instead of canvas
