+++ ui/src/features/Canvas/components/index.ts
@@ -1,2 +1,1 @@
 export { default as CanvasContextMenu } from "./CanvasContextMenu";
-export { default as NodeDeletionDialog } from "./NodeDeletionDialog";
