+++ ui/src/features/Canvas/hooks.ts
@@ -74,12 +74,6 @@ export default ({
     onEdgesChange,
   });
 
-  const [showBeforeDeleteDialog, setShowBeforeDeleteDialog] =
-    useState<boolean>(false);
-  const deferredDeleteRef = useRef<{
-    resolve: (val: boolean) => void;
-  } | null>(null);
-
   const [contextMenu, setContextMenu] = useState<ContextMenuMeta | null>(null);
   const paneRef = useRef<HTMLDivElement>(null);
   const getContextMenuPosition = (event: MouseEvent) => {
@@ -146,50 +140,6 @@ export default ({
     [setContextMenu],
   );
 
-  const handleBeforeDeleteNodes = useCallback(
-    ({ nodes: nodesToDelete }: { nodes: Node[] }) => {
-      return new Promise<boolean>((resolve) => {
-        const deletingIds = new Set(nodesToDelete.map((node) => node.id));
-
-        let totalInputRouters = 0;
-        let totalOutputRouters = 0;
-        let remainingInputRouters = 0;
-        let remainingOutputRouters = 0;
-
-        for (const node of nodes) {
-          const officalName = node.data.officialName;
-          if (officalName !== "InputRouter" && officalName !== "OutputRouter")
-            continue;
-          const isDeleting = deletingIds.has(node.id);
-
-          if (officalName === "InputRouter") {
-            totalInputRouters++;
-            if (!isDeleting) remainingInputRouters++;
-          } else if (officalName === "OutputRouter") {
-            totalOutputRouters++;
-            if (!isDeleting) remainingOutputRouters++;
-          }
-        }
-
-        const isDeletingLastInputRouter =
-          totalInputRouters > 0 && remainingInputRouters === 0;
-
-        const isDeletingLastOutputRouter =
-          totalOutputRouters > 0 && remainingOutputRouters === 0;
-
-        if (isDeletingLastInputRouter || isDeletingLastOutputRouter) {
-          deferredDeleteRef.current = { resolve };
-          setShowBeforeDeleteDialog(true);
-        } else {
-          resolve(true);
-        }
-      });
-    },
-    [nodes],
-  );
-
-  const handleDeleteDialogClose = () => setShowBeforeDeleteDialog(false);
-
   const handleCloseContextmenu = () => {
     setContextMenu(null);
   };
@@ -197,7 +147,6 @@ export default ({
   return {
     handleNodesChange,
     handleNodesDelete,
-    handleBeforeDeleteNodes,
     handleNodeDragStop,
     handleNodeDragOver,
     handleNodeDrop,
@@ -209,10 +158,7 @@ export default ({
     handleSelectionContextMenu,
     handlePaneContextMenu,
     handleCloseContextmenu,
-    handleDeleteDialogClose,
     contextMenu,
     paneRef,
-    showBeforeDeleteDialog,
-    deferredDeleteRef,
   };
 };
