+++ ui/src/features/Editor/components/index.ts
@@ -4,3 +4,4 @@ export { default as WorkflowTabs } from "./WorkflowTabs";
 export { default as LeftPanel } from "./LeftPanel";
 export { default as ParamsPanel } from "./ParamsPanel";
 export { default as RightPanel } from "./RightPanel";
+export { default as NodeDeletionDialog } from "./NodeDeletionDialog";
