+++ ui/src/features/Editor/hooks.ts
@@ -77,6 +77,11 @@ export default ({
     setSelectedEdgeIds,
     undoTrackerActionWrapper,
   });
+  const [showBeforeDeleteDialog, setShowBeforeDeleteDialog] =
+    useState<boolean>(false);
+  const deferredDeleteRef = useRef<{
+    resolve: (val: boolean) => void;
+  } | null>(null);
 
   const isSubworkflow = useMemo(() => {
     if (!currentYWorkflow) return false;
@@ -226,6 +231,50 @@ export default ({
     rawWorkflows,
   });
 
+  const handleBeforeDeleteNodes = useCallback(
+    ({ nodes: nodesToDelete }: { nodes: Node[] }) => {
+      return new Promise<boolean>((resolve) => {
+        const deletingIds = new Set(nodesToDelete.map((node) => node.id));
+
+        let totalInputRouters = 0;
+        let totalOutputRouters = 0;
+        let remainingInputRouters = 0;
+        let remainingOutputRouters = 0;
+
+        for (const node of nodes) {
+          const officalName = node.data.officialName;
+          if (officalName !== "InputRouter" && officalName !== "OutputRouter")
+            continue;
+          const isDeleting = deletingIds.has(node.id);
+
+          if (officalName === "InputRouter") {
+            totalInputRouters++;
+            if (!isDeleting) remainingInputRouters++;
+          } else if (officalName === "OutputRouter") {
+            totalOutputRouters++;
+            if (!isDeleting) remainingOutputRouters++;
+          }
+        }
+
+        const isDeletingLastInputRouter =
+          totalInputRouters > 0 && remainingInputRouters === 0;
+
+        const isDeletingLastOutputRouter =
+          totalOutputRouters > 0 && remainingOutputRouters === 0;
+
+        if (isDeletingLastInputRouter || isDeletingLastOutputRouter) {
+          deferredDeleteRef.current = { resolve };
+          setShowBeforeDeleteDialog(true);
+        } else {
+          resolve(true);
+        }
+      });
+    },
+    [nodes],
+  );
+
+  const handleDeleteDialogClose = () => setShowBeforeDeleteDialog(false);
+
   useShortcuts([
     {
       keyBinding: { key: "r", commandKey: false },
@@ -297,6 +346,8 @@ export default ({
     canRedo,
     isMainWorkflow,
     hasReader,
+    deferredDeleteRef,
+    showBeforeDeleteDialog,
     handleRightPanelOpen,
     handleWorkflowAdd: handleYWorkflowAdd,
     handleWorkflowDeployment,
@@ -310,6 +361,8 @@ export default ({
     handleLayoutChange,
     handleNodesAdd: handleYNodesAdd,
     handleNodesChange: handleYNodesChange,
+    handleBeforeDeleteNodes,
+    handleDeleteDialogClose,
     handleNodeHover,
     handleNodeDataUpdate: handleYNodeDataUpdate,
     handleNodeSettings,
